[Feature] getField: Add get_field proxy function (…)
Add a getField as a first proxy function to reuse template in the frontend and backend
Essentially this allows to use Timber features by abstracting them away behind a function. This function can then be implemented or mocked in Twig.js

To enable this I needed to reimplement inline includes (credits to @backflip's gulp-twig) in a fork of @zimmen's gulp-twig because the later allows for custom twig functions, filters and tags and is actively maintained.